Flask/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
redis==4.1.1
rq==1.9.0
Flask-Cors==3.0.10
Pillow==7.1.2
firebase-admin==5.0.1